Update model init in factory func. Move models to src directory.
This is so that there is no ambiguity in where the models will be
stored when installed by a user, regardless of OS. This makes the
repo structure a bit less than ideal but the models will now be
easily accessible as package data to other submodules/files.

This discussion was helpful in highlighting these pros/cons:
https://stackoverflow.com/questions/4519127/setuptools-package-data-folder-location

Update models dir in setup.cfg. Update factory func for dir change.

Untrack models dir